Ensure that we don't try to delete the importable when it doesn't exist

+++ app/models/nfg_csv_importer/imported_record.rb
@@ -14,15 +14,19 @@ class NfgCsvImporter::ImportedRecord < ActiveRecord::Base
   end
 
   def destroy
-    if importable.respond_to?(:can_be_destroyed?)
-      if importable.can_be_destroyed?
+    if self.importable.present?
+      if importable.respond_to?(:can_be_destroyed?)
+        if importable.can_be_destroyed?
+          importable.destroy
+          super
+        else
+          return
+        end
+      else
         importable.destroy
         super
-      else
-        return
       end
     else
-      importable.destroy
       super
     end
   end
